packages/common/pyproject.toml
[tool.poetry]
name = "miose-toolkit-common"
version = "0.2.0"
description = "Miose Toolkit 通用模块"
authors = ["KroMiose <li_xiangff@163.com>"]
packages = [{ include = "miose_toolkit_common", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.8, <3.12"
aiohttp = "^3.9.2"
requests = "^2.31.0"
poetry = "^1.6.1"
pyyaml = "^6.0.1"
cryptography = ">=42.0.4"
pydantic = ">=1.10, <3.0"

[tool.poetry.scripts]
test = "src.tests:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
